+++ renku/ui/service/serializers/workflows.py
@@ -22,11 +22,11 @@ from marshmallow_oneofschema import OneOfSchema
 
 from renku.domain_model.dataset import DatasetCreatorsJson
 from renku.ui.cli.utils.plugins import get_supported_formats
-from renku.ui.service.serializers.common import LocalRepositorySchema, RemoteRepositorySchema
+from renku.ui.service.serializers.common import RemoteRepositorySchema
 from renku.ui.service.serializers.rpc import JsonRPCResponse
 
 
-class WorkflowPlansListRequest(LocalRepositorySchema, RemoteRepositorySchema):
+class WorkflowPlansListRequest(RemoteRepositorySchema):
     """Request schema for plan list view."""
 
 
@@ -64,7 +64,7 @@ class WorkflowPlansListResponseRPC(JsonRPCResponse):
     result = fields.Nested(WorkflowPlansListResponse)
 
 
-class WorkflowPlansShowRequest(LocalRepositorySchema, RemoteRepositorySchema):
+class WorkflowPlansShowRequest(RemoteRepositorySchema):
     """Request schema for plan show view."""
 
     plan_id = fields.String(required=True)
@@ -201,7 +201,7 @@ WorkflowExportFormatEnum = Enum(  # type: ignore
 )
 
 
-class WorkflowPlansExportRequest(LocalRepositorySchema, RemoteRepositorySchema):
+class WorkflowPlansExportRequest(RemoteRepositorySchema):
     """Request schema for exporting a plan."""
 
     plan_id = fields.String(required=True)