+++ CHANGES.rst
@@ -18,6 +18,36 @@
 Changes
 =======
 
+`2.7.0 <https://github.com/SwissDataScienceCenter/renku-python/compare/v2.6.1...v2.7.0>`__ (2023-08-09)
+-------------------------------------------------------------------------------------------------------
+
+Bug Fixes
+~~~~~~~~~
+
+-  **cli:** adapt to changes in KG api for importing datasets
+   (`#3549 <https://github.com/SwissDataScienceCenter/renku-python/issues/3549>`__)
+   (`020434a <https://github.com/SwissDataScienceCenter/renku-python/commit/020434a7dd6449755644a2e9ca849b8821900f72>`__)
+-  **cli:** fix overwriting dataset tags
+   (`#3497 <https://github.com/SwissDataScienceCenter/renku-python/issues/3497>`__)
+   (`17a49ae <https://github.com/SwissDataScienceCenter/renku-python/commit/17a49aebc7207901986c38fd30afc58baf40be1f>`__)
+-  **service:** add branch to service cache path
+   (`#3562 <https://github.com/SwissDataScienceCenter/renku-python/issues/3562>`__)
+   (`3800a38 <https://github.com/SwissDataScienceCenter/renku-python/commit/3800a3823515763c207b1b15f348df3b0cdd9831>`__)
+-  **service:** add support for using default values in template
+   parameters
+   (`#3550 <https://github.com/SwissDataScienceCenter/renku-python/issues/3550>`__)
+   (`d162392 <https://github.com/SwissDataScienceCenter/renku-python/commit/d162392b3dc20dd3433be78b08f101e7f268ed7d>`__)
+
+Features
+~~~~~~~~
+
+-  **cli:** allow disabling automated parameter detection in renku run
+   (`#3548 <https://github.com/SwissDataScienceCenter/renku-python/issues/3548>`__)
+   (`bcdeba1 <https://github.com/SwissDataScienceCenter/renku-python/commit/bcdeba1f286ad786edb4e19e21236c4408fa3ed7>`__)
+-  **service:** replace/refactor internal repository cache
+   (`#3534 <https://github.com/SwissDataScienceCenter/renku-python/issues/3534>`__)
+   (`a9994a8 <https://github.com/SwissDataScienceCenter/renku-python/commit/a9994a8cb2541527ca9be731037cc60b03d62cea>`__)
+
 `2.6.0 <https://github.com/SwissDataScienceCenter/renku-python/compare/v2.5.0...v2.6.0>`__ (2023-06-20)
 -------------------------------------------------------------------------------------------------------